Tweak import script

+++ bin/import_production_database.sh
@@ -1,11 +1,14 @@
 #!/bin/bash
 
-# Ensure we are starting fresh
-[ -e latest.dump ] && rm latest.dump
+# Check if USE_EXISTING_DUMP is set to true
+if [ "$USE_EXISTING_DUMP" != "true" ]; then
+  # Ensure we are starting fresh
+  [ -e latest.dump ] && rm latest.dump
 
-# Capture and download the backup from Heroku (only if you haven't already done this)
-heroku pg:backups:capture -a echosight-production
-heroku pg:backups:download -a echosight-production
+  # Capture and download the backup from Heroku (only if you haven't already done this)
+  heroku pg:backups:capture -a echosight-production
+  heroku pg:backups:download -a echosight-production
+fi
 
 # Drop and recreate the development database
 DISABLE_DATABASE_ENVIRONMENT_CHECK=1 rails db:drop
